Adapt module and files layout for plugin implementation
This PR is separated and prepared for #428 to avoid mixing changes, especially for moving
modules/files without changing any implementations. See #428 (comment)
for the design of restructuring the modules.
It include following changes:

- Move strucute selction to step1 module
- Move modules/folders to configuration folder (step2)
- Move modules/folders to submission (step3)
- Move modules/folders to results (steps)
- move modules/widgets to common folder
- load registred viewer in __init__ to trigger
