+++ src/aiidalab_qe/__main__.py
@@ -3,9 +3,9 @@
 
 import click
 
-from aiidalab_qe.app.setup_codes import codes_are_setup
-from aiidalab_qe.app.setup_codes import install as install_qe_codes
-from aiidalab_qe.app.sssp import install as setup_sssp
+from aiidalab_qe.app.common.setup_codes import codes_are_setup
+from aiidalab_qe.app.common.setup_codes import install as install_qe_codes
+from aiidalab_qe.app.submission.sssp import install as setup_sssp
 
 
 @click.group()
