+++ src/aiidalab_qe/app/configuration/workflow.py
@@ -1,0 +1,158 @@
+# -*- coding: utf-8 -*-
+"""Widgets for the submission of bands work chains.
+
+Authors: AiiDAlab team
+"""
+import ipywidgets as ipw
+
+from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
+
+
+class WorkChainSettings(ipw.VBox):
+    structure_title = ipw.HTML(
+        """<div style="padding-top: 0px; padding-bottom: 0px">
+        <h4>Structure</h4></div>"""
+    )
+    structure_help = ipw.HTML(
+        """<div style="line-height: 140%; padding-top: 0px; padding-bottom: 5px">
+        You have three options:<br>
+        (1) Structure as is: perform a self consistent calculation using the structure provided as input.<br>
+        (2) Atomic positions: perform a full relaxation of the internal atomic coordinates. <br>
+        (3) Full geometry: perform a full relaxation for both the internal atomic coordinates and the cell vectors. </div>"""
+    )
+    materials_help = ipw.HTML(
+        """<div style="line-height: 140%; padding-top: 10px; padding-bottom: 10px">
+        Below you can indicate both if the material should be treated as an insulator
+        or a metal (if in doubt, choose "Metal"),
+        and if it should be studied with magnetization/spin polarization,
+        switch magnetism On or Off (On is at least twice more costly).
+        </div>"""
+    )
+
+    properties_title = ipw.HTML(
+        """<div style="padding-top: 0px; padding-bottom: 0px">
+        <h4>Properties</h4></div>"""
+    )
+    properties_help = ipw.HTML(
+        """<div style="line-height: 140%; padding-top: 10px; padding-bottom: 0px">
+        The band structure workflow will
+        automatically detect the default path in reciprocal space using the
+        <a href="https://www.materialscloud.org/work/tools/seekpath" target="_blank">
+        SeeK-path tool</a>.</div>"""
+    )
+
+    protocol_title = ipw.HTML(
+        """<div style="padding-top: 0px; padding-bottom: 0px">
+        <h4>Protocol</h4></div>"""
+    )
+    protocol_help = ipw.HTML(
+        """<div style="line-height: 140%; padding-top: 6px; padding-bottom: 0px">
+        The "moderate" protocol represents a trade-off between
+        accuracy and speed. Choose the "fast" protocol for a faster calculation
+        with less precision and the "precise" protocol to aim at best accuracy (at the price of longer/costlier calculations).</div>"""
+    )
+
+    def __init__(self, **kwargs):
+        # RelaxType: degrees of freedom in geometry optimization
+        self.relax_type = ipw.ToggleButtons(
+            options=[
+                ("Structure as is", "none"),
+                ("Atomic positions", "positions"),
+                ("Full geometry", "positions_cell"),
+            ],
+            value="positions_cell",
+        )
+
+        # SpinType: magnetic properties of material
+        self.spin_type = ipw.ToggleButtons(
+            options=[("Off", "none"), ("On", "collinear")],
+            value=DEFAULT_PARAMETERS["spin_type"],
+            style={"description_width": "initial"},
+        )
+
+        # ElectronicType: electronic properties of material
+        self.electronic_type = ipw.ToggleButtons(
+            options=[("Metal", "metal"), ("Insulator", "insulator")],
+            value=DEFAULT_PARAMETERS["electronic_type"],
+            style={"description_width": "initial"},
+        )
+
+        # Checkbox to see if the band structure should be calculated
+        self.bands_run = ipw.Checkbox(
+            description="",
+            indent=False,
+            value=True,
+            layout=ipw.Layout(max_width="10%"),
+        )
+
+        # Checkbox to see if the PDOS should be calculated
+        self.pdos_run = ipw.Checkbox(
+            description="",
+            indent=False,
+            value=True,
+            layout=ipw.Layout(max_width="10%"),
+        )
+
+        # Work chain protocol
+        self.workchain_protocol = ipw.ToggleButtons(
+            options=["fast", "moderate", "precise"],
+            value="moderate",
+        )
+        super().__init__(
+            children=[
+                self.structure_title,
+                self.structure_help,
+                self.relax_type,
+                self.materials_help,
+                ipw.HBox(
+                    children=[
+                        ipw.Label(
+                            "Electronic Type:",
+                            layout=ipw.Layout(
+                                justify_content="flex-start", width="120px"
+                            ),
+                        ),
+                        self.electronic_type,
+                    ]
+                ),
+                ipw.HBox(
+                    children=[
+                        ipw.Label(
+                            "Magnetism:",
+                            layout=ipw.Layout(
+                                justify_content="flex-start", width="120px"
+                            ),
+                        ),
+                        self.spin_type,
+                    ]
+                ),
+                self.properties_title,
+                ipw.HTML("Select which properties to calculate:"),
+                ipw.HBox(children=[ipw.HTML("<b>Band structure</b>"), self.bands_run]),
+                ipw.HBox(
+                    children=[
+                        ipw.HTML("<b>Projected density of states</b>"),
+                        self.pdos_run,
+                    ]
+                ),
+                self.properties_help,
+                self.protocol_title,
+                ipw.HTML("Select the protocol:", layout=ipw.Layout(flex="1 1 auto")),
+                self.workchain_protocol,
+                self.protocol_help,
+            ],
+            **kwargs,
+        )
+
+    def _update_settings(self, **kwargs):
+        """Update the settings based on the given dict."""
+        for key in [
+            "relax_type",
+            "spin_type",
+            "electronic_type",
+            "bands_run",
+            "pdos_run",
+            "workchain_protocol",
+        ]:
+            if key in kwargs:
+                getattr(self, key).value = kwargs[key]
