+++ src/aiidalab_qe/app/submission/resource.py
@@ -1,0 +1,85 @@
+# -*- coding: utf-8 -*-
+"""Widgets for the submission of bands work chains.
+
+Authors: AiiDAlab team
+"""
+import ipywidgets as ipw
+
+
+class ResourceSelectionWidget(ipw.VBox):
+    """Widget for the selection of compute resources."""
+
+    title = ipw.HTML(
+        """<div style="padding-top: 0px; padding-bottom: 0px">
+        <h4>Resources</h4>
+    </div>"""
+    )
+    prompt = ipw.HTML(
+        """<div style="line-height:120%; padding-top:0px">
+        <p style="padding-bottom:10px">
+        Specify the resources to use for the pw.x calculation.
+        </p></div>"""
+    )
+
+    def __init__(self, **kwargs):
+        extra = {
+            "style": {"description_width": "150px"},
+            "layout": {"min_width": "180px"},
+        }
+        self.num_nodes = ipw.BoundedIntText(
+            value=1, step=1, min=1, max=1000, description="Nodes", **extra
+        )
+        self.num_cpus = ipw.BoundedIntText(
+            value=1, step=1, min=1, description="CPUs", **extra
+        )
+
+        super().__init__(
+            children=[
+                self.title,
+                ipw.HBox(
+                    children=[self.prompt, self.num_nodes, self.num_cpus],
+                    layout=ipw.Layout(justify_content="space-between"),
+                ),
+            ]
+        )
+
+    def reset(self):
+        self.num_nodes.value = 1
+        self.num_cpus.value = 1
+
+
+class ParallelizationSettings(ipw.VBox):
+    """Widget for setting the parallelization settings."""
+
+    title = ipw.HTML(
+        """<div style="padding-top: 0px; padding-bottom: 0px">
+        <h4>Parallelization</h4>
+    </div>"""
+    )
+    prompt = ipw.HTML(
+        """<div style="line-height:120%; padding-top:0px">
+        <p style="padding-bottom:10px">
+        Specify the number of k-points pools for the calculations.
+        </p></div>"""
+    )
+
+    def __init__(self, **kwargs):
+        extra = {
+            "style": {"description_width": "150px"},
+            "layout": {"min_width": "180px"},
+        }
+        self.npools = ipw.BoundedIntText(
+            value=1, step=1, min=1, max=128, description="Number of k-pools", **extra
+        )
+        super().__init__(
+            children=[
+                self.title,
+                ipw.HBox(
+                    children=[self.prompt, self.npools],
+                    layout=ipw.Layout(justify_content="space-between"),
+                ),
+            ]
+        )
+
+    def reset(self):
+        self.npools.value = 1
