+++ tests/test_pseudo.py
@@ -1,6 +1,6 @@
 def test_pseudos_family_selector_widget():
     """Test the pseudos widget."""
-    from aiidalab_qe.app.pseudos import PseudoFamilySelector
+    from aiidalab_qe.app.configuration.pseudos import PseudoFamilySelector
 
     wg = PseudoFamilySelector()
     wg.override_protocol_pseudo_family.value = True
